pyproject.toml
[project]
name = "cake-airflow-sftp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.11"
dependencies = [
    "apache-airflow==3.0.4",
    "apache-airflow-providers-sftp>=5.3.4",
    "ruff>=0.12.8",
]
